pyproject.toml
[project]
name = "fastapi-keycloak"
version = "0.1.0"
description = "Projeto feito como template de integração entre as ferramentas FastAPI e Keycloak"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "python-jose>=3.5.0",
    "requests>=2.32.5",
    "uvicorn>=0.42.0",
]
